Minor refactor of files in test directory
